fix: Allow to create new game while already in a game session
Previously, a new game map would be generated immediately upon starting
up the game, and choosing New Game from the main menu would simply move
you to the map.

This meant that if you loaded or started a game, returned to the main
menu, and _then_ selected New Game, it would have the same effect as
clicking Load Game, booting you back to the game in progress.

Now the map generation is done when you click New Game, meaning that:
  a) the game doesn't waste time generating a new map if you're just going to
load a saved game, and
  b) If you choose New Game in the middle of a session, it will
correctly start a new game